+++ Sources/OpenJoystickDriver/CLI.swift
@@ -22,7 +22,7 @@ struct CLI {
     case "uninstall": UninstallCommand().run()
     case "run": RunCommand().run()
     case "--help", "-h", "help": printHelp()
-    case "--version", "-v", "version": print("OpenJoystickDriver v0.5.0-alpha.4")
+    case "--version", "-v", "version": print("OpenJoystickDriver v0.5.0-alpha.5")
     default:
       print("Unknown command: \(command)")
       printHelp()
@@ -33,7 +33,7 @@ struct CLI {
   private func printHelp() {
     print(
       """
-      OpenJoystickDriver v0.5.0-alpha.4 \
+      OpenJoystickDriver v0.5.0-alpha.5 \
       - macOS gamepad driver
 
       Usage: OpenJoystickDriver \